### 2023-11-04 #6019 @editorialbot
[PRE REVIEW]: Enlsip.jl: A Julia optimization package to solve constrained nonlinear least-squares problems
**Submitting author:** @pierre-borie (<a href="http://orcid.org/0009-0000-1043-5057">Pierre Borie</a>)
**Repository:** https://github.com/UncertainLab/Enlsip.jl.git
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.9.1
