### 2024-01-08 #6019 @jbytecode
👋👋👋 Dear  @matbesancon, @anriseth,  @tmigot, @lootie, @odunbar 👋👋👋

**Enlsip.jl: A Julia optimization package to solve constrained nonlinear least-squares problems**

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/6019).

Thank you in advance!
