### 2024-02-02 UncertainLab/Enlsip.jl#3 @tmigot
[JOSS review] typos in the paper
- [x] Typos:
   - [x] At least once ENLSIP was noted ENLISP.
   - [x] "It implements a nonlinear least squares under nonlinear constraints solver" -> "It implements a solver for nonlinear least squares with nonlinear constraints"

- [x] * Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?* I think the theoretical guarantees offer by the solver should be clarified for a general audience. For instance, I assume this is not a global optimizer. In general, the paper should target a broad audience.
- [x] I think the paper lacks a brief description of the algorithm. It is not clear if this uses the second-order derivatives of the involved function or if the active-set strategy used in the paper is exactly the same as in the book by Nocedal & Wright ...
- [x] *Data sharing: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.* A large part of the discussion in the paper is comparing the Fortran version vs. the Julia version while there is no shared data on this.
- [x] This recent post could help the author expand the list of existing solvers for (unconstrained) NLS **https://discourse.julialang.org/t/comparing-non-linear-least-squares-solvers/104752**.
- [x] It would also help states why we consider the NLS case separately to the classical optimization case.

https://github.com/openjournals/joss-reviews/issues/6226
