### 2024-02-19 UncertainLab/Enlsip.jl#2 @pierre-borie
First of all, please accept my apologies for taking this much time to respond your review comments.

To answer your General Checks remarks, the other authors did not actually contribute to the code per se but they were important stakeholders in the conception of this package which is why we thought relevant to refer them as authors. To bring some details, Alain Marcotte helped me to carry out the comparison tests between Julia and Fortran and the other two supervised the conduct of this project.
As you already kind of guessed, the code was first developed on a local repository and then transfered to my lab GitHub when we published it on the Julia package manager, which explains the small number of commits and the current versioning. I hope this clarifies the background of the package!

Thanks a lot for your suggestions of improvement. I started to bring corrections and additions to the package source code and documentation in accordance with your remarks and will come back to you once I finished completing them as you suggested.

I am, once again, sorry for my response delay.
