### 2024-02-23 759a7bb
Expanding default number type form Float64 to AbstractFloat
