+++ src/enlsip_functions.jl
@@ -14,7 +14,7 @@ Parameters :
     the number of rows of matrix `T`.
 =#
 
-function pseudo_rank(diag_T::Vector{Float64}, ε_rank::Float64)
+function pseudo_rank(diag_T::Vector{<:AbstractFloat}, ε_rank::AbstractFloat)
 
     if isempty(diag_T) || abs(diag_T[1]) < ε_rank
         pseudo_rank = 0
@@ -114,18 +114,18 @@ Finally, the search direction is computed by forming : `p = Q1 * [p1 ; P3*p2]`
 * `d` : vector of size `m`, contains the right handside of the system solved to compute `p2`
 =# 
 function sub_search_direction(
-    J1::Matrix{Float64},
-    rx::Vector{Float64},
-    cx::Vector{Float64},
+    J1::Matrix{<:AbstractFloat},
+    rx::Vector{<:AbstractFloat},
+    cx::Vector{<:AbstractFloat},
     F_A::Factorization,
     F_L11::Factorization,
     F_J2::Factorization,
-    n::Int64,
-    t::Int64,
-    rankA::Int64,
-    dimA::Int64,
-    dimJ2::Int64,
-    code::Int64)
+    n::Int,
+    t::Int,
+    rankA::Int,
+    dimA::Int,
+    dimJ2::Int,
+    code::Int)
 
     # Solving without stabilization 
     if code == 1
@@ -204,14 +204,14 @@ If `rankA = t`, the first system is solved, otherwise, the second one is solved.
 * `F_J2` : QR decomposition of Matrix `J2` defined in [`sub_search_direction`](@ref)
 =#
 function gn_search_direction(
-    J::Matrix{Float64},
-    rx::Vector{Float64},
-    cx::Vector{Float64},
+    J::Matrix{<:AbstractFloat},
+    rx::Vector{<:AbstractFloat},
+    cx::Vector{<:AbstractFloat},
     F_A::Factorization,
     F_L11::Factorization,
-    rankA::Int64,
-    t::Int64,
-    ε_rank::Float64,
+    rankA::Int,
+    t::Int,
+    ε_rank::AbstractFloat,
     current_iter::Iteration)
 
     code = (rankA == t ? 1 : -1)
@@ -242,11 +242,11 @@ end
 
 function hessian_res!(
     r::ResidualsFunction,
-    x::Vector{Float64},
-    rx::Vector{Float64},
-    n::Int64,
-    m::Int64,
-    B::Matrix{Float64})
+    x::Vector{<:AbstractFloat},
+    rx::Vector{<:AbstractFloat},
+    n::Int,
+    m::Int,
+    B::Matrix{<:AbstractFloat})
 
     # Only residuals evaluation
     r.ctrl = 1
@@ -287,13 +287,13 @@ end
 
 function hessian_cons!(
     c::ConstraintsFunction,
-    x::Vector{Float64},
-    λ::Vector{Float64},
-    active::Vector{Int64},
-    n::Int64,
-    l::Int64,
-    t::Int64,
-    B::Matrix{Float64})
+    x::Vector{<:AbstractFloat},
+    λ::Vector{<:AbstractFloat},
+    active::Vector{<:Int},
+    n::Int,
+    l::Int,
+    t::Int,
+    B::Matrix{<:AbstractFloat})
 
     # Only constraints evaluation
     c.ctrl = 1
@@ -348,17 +348,17 @@ end
 
 
 function newton_search_direction(
-    x::Vector{Float64},
+    x::Vector{<:AbstractFloat},
     c::ConstraintsFunction,
     r::ResidualsFunction,
-    active_cx::Vector{Float64},
[…]
