+++ docs/src/index.md
@@ -1,27 +1,27 @@
 # [Enlsip.jl documentation](@id Home)
 
-
 ## Introduction
 
 Package `Enlsip.jl` is the Julia version of an eponymous Fortran77 library (ENLSIP standing for Easy Nonlinear Least Squares Inequalities Program) designed to solve nonlinear least squares problems under nonlinear constraints.
 
-The optimization method implemented in Enlsip was conceived in the late 1980s by two swedish authors named Per Lindström and Per Åke Wedin [^1].
+The optimization method implemented in `Enlsip.jl` was conceived in the late 1980s by two swedish authors named Per Lindström and Per Åke Wedin [^1]. 
 
-Problems that can be solved using Enlsip are modeled as follows:
+It is designed for solve nonlinear least squares problems subject to (s.t.) nonlinear constraints, which can be modeled as the following optimization problem:
 
 ```math
 \begin{aligned}
-\min \quad &  \dfrac{1}{2} \|r(x)\|^2 \\
-\text{s.t.} \quad & c_i(x) = 0, \quad i =1,\ldots,q \\
-& c_j(x) \geq 0, \quad j=q+1,\ldots,\ell, \\
+\min_{x \in \mathbb{R}^n} \quad &  \dfrac{1}{2} \|r(x)\|^2 \\
+\text{s.t.} \quad & c_i(x) = 0, \quad i \in \mathcal{E} \\
+& c_i(x) \geq 0, \quad i \in \mathcal{I}, \\
 \end{aligned}
 ```
 
-where $r:\mathbb{R}^n\rightarrow\mathbb{R}^m$, the residuals, and
-$c:\mathbb{R}^n\rightarrow\mathbb{R}^{q+\ell}$, concatenation of the constraints, are $\mathcal{C}^1$ multi-functions.
+where:
+
+* the residuals $r_i:\mathbb{R}^n\rightarrow\mathbb{R}$ and the constraints $c_i:\mathbb{R}^n\rightarrow\mathbb{R}$ are assumed to be $\mathcal{C}^2$ functions;
+* norm $\|\cdot\|$ denotes the Euclidean norm.
 
 Note that box constraints are modeled as general inequality constraints.
-
 
 ## How to install
 
@@ -34,13 +34,35 @@ Pkg.add("Enlsip")
 
 ## How to use
 
-Using `Enlsip.jl` to solve optimization problems consists in, first, instantiating a model and then call the solver on it. 
+Using `Enlsip.jl` to solve optimization problems consists in, first, instantiating a model and then call the solver on it.
 
 Details and examples with problems from the litterature in the [Usage](@ref) page.
 
-## Contents
+## Description of the algorithm
 
-```@contents
+Starting from a point $x_0$, the algorithm builds a sequence $(x_k)_k$ converging to a first order critical point of the problem.
+
+At a given iteration $k$ , a search direction $p_k\in\mathbb{R}^n$ and a steplength $\alpha_k\in[0,1]$ are computed such that the next step is $x_{k+1}:=x_k+\alpha_kp_k$.
+
+The search direction is the solution of a linear least squares problem under linear constraints.
+
+The objective of the subproblem is obtained after linearizing the residuals in a small neighborhood of current point $x_k$ , resulting into the Gauss-Newton approximation: 
+$$r(x_k+p)\approx J_kp+r_k,$$
+
+where $J_k$ denotes the jacobian matrix of the residuals function evaluated at $x_k$ and $r_k:=r(x_k)$ .
+
[…]
