### 2024-02-27 d4a1ae4
new docstrig for status and community guidelines
