+++ docs/src/index.md
@@ -4,7 +4,7 @@
 
 Package `Enlsip.jl` is the Julia version of an eponymous Fortran77 library (ENLSIP standing for Easy Nonlinear Least Squares Inequalities Program) designed to solve nonlinear least squares problems under nonlinear constraints.
 
-The optimization method implemented in `Enlsip.jl` was conceived in the late 1980s by two swedish authors named Per Lindström and Per Åke Wedin [^1]. 
+The optimization method implemented in `Enlsip.jl` was conceived in the early 1980s by two swedish authors named Per Lindström and Per Åke Wedin [^1].
 
 It is designed for solve nonlinear least squares problems subject to (s.t.) nonlinear constraints, which can be modeled as the following optimization problem:
 
@@ -36,7 +36,7 @@ Pkg.add("Enlsip")
 
 Using `Enlsip.jl` to solve optimization problems consists in, first, instantiating a model and then call the solver on it.
 
-Details and examples with problems from the litterature in the [Usage](@ref) page.
+Details and examples with problems from the literature in the [Usage](@ref) page.
 
 ## Description of the algorithm
 
@@ -44,16 +44,19 @@ Starting from a point $x_0$, the algorithm builds a sequence $(x_k)_k$ convergin
 
 At a given iteration $k$ , a search direction $p_k\in\mathbb{R}^n$ and a steplength $\alpha_k\in[0,1]$ are computed such that the next step is $x_{k+1}:=x_k+\alpha_kp_k$.
 
-The search direction is the solution of a linear least squares problem under linear constraints.
+### Search direction
 
-The objective of the subproblem is obtained after linearizing the residuals in a small neighborhood of current point $x_k$ , resulting into the Gauss-Newton approximation: 
+The search direction is the solution of a subproblem derived from an approximation of the original problem.
+
+First, the residuals are linearized in a small neighborhood of current point $x_k$:
+
 $$r(x_k+p)\approx J_kp+r_k,$$
 
-where $J_k$ denotes the jacobian matrix of the residuals function evaluated at $x_k$ and $r_k:=r(x_k)$ .
+where $J_k$ denotes the Jacobian matrix of the residuals function evaluated at $x_k$ and $r_k:=r(x_k)$ . The resulting objective function of the subproblem corresponds to the Gauss-Newton approximation.
 
-The constraints of the subproblem are formed after a subset of the constraints of the original problem, often denoted as the working set in the literature. It contains all the equality constraints but only the inequality constraints that are believed to be satisfied with equality at the solution. This subset, updated at every iteration, can thus be seen as a guess of the optimal active set[^2]. 
+The constraints of the subproblem are formed after a subset of the constraints of the original problem. It contains all the equality constraints but only the inequality constraints that are believed to be satisfied with equality at the solution. This subset, often denoted as the working set in the literature, is updated at every iteration and can be seen as a guess of the optimal active set[^2].
 
-Noting $\hat{c}_k$ the vector of constraints in the working set evaluated at $x_k$ and $\hat{A}_k$ its jacobian matrix, the subproblem is then given by:
+Noting $\hat{c}_k$ the vector of constraints in the working set evaluated at $x_k$ and $\hat{A}_k$ its Jacobian matrix, the subproblem is then given by:
 
 ```math
 \begin{aligned}
@@ -62,7 +65,34 @@ Noting $\hat{c}_k$ the vector of constraints in the working set evaluated at $x_
 \end{aligned}
 ```
 
+The subproblem is then solved by a null-space type method.
+
+### Steplength
+
[…]
