+++ docs/src/tutorial.md
@@ -1,11 +1,12 @@
-```@meta 
+# [Usage](@id Usage)
+
+```@meta
 CurrentModule = Enlsip
 ```
 
 ```@setup tutorial
 using Enlsip
 ```
-# [Usage](@id Usage)
 
 This sections provides details on how to instantiate and solve a constrained least squares problem with `Enlsip.jl`
 As a reminder from [Home](@ref), problems to solve are of the following form:
@@ -18,9 +19,13 @@ As a reminder from [Home](@ref), problems to solve are of the following form:
 \end{aligned}
 ```
 
+with:
+
+* the residuals $r_i:\mathbb{R}^n\rightarrow\mathbb{R}$ and the constraints $c_i:\mathbb{R}^n\rightarrow\mathbb{R}$ assumed to be $\mathcal{C}^2$ functions;
+* norm $\|\cdot\|$ denoting the Euclidean norm.
 Note that with this formulation, bounds constraints are not distinguished from general inequality constraints. Though, as shown later in this section, they can be provided as vectors of lower and/or upper bounds, which is more convenient for this type of constraints.
 
-Also, the `Enlsip` solver works with double precision float numbers (i.e. type `Float64`).
+It should be borne in mind, however, that the method implemented in `Enlsip.jl` has been conceived for nonlinear problems, as there is no other assumption made about the nature of the residuals and constraints functions, apart from being two-time continously differentiable. The algorithm can still be used to solve linear least squares subject to linear constraints but it will not be as effective as other software where those aspects are taken into account in the design of the optimization method.
 
 ## Instantiate a model
 
@@ -28,8 +33,7 @@ Solving a problem with Enlsip is organized in two steps.
 
 First, a model of type [`CnlsModel`](@ref) must be instantiated.
 
-
-The `CnlsModel` constructor requires the evaluation functions of residuals, constraints, their associated jacobian matrices and dimensions of the problem. 
+The `CnlsModel` constructor requires the evaluation functions of residuals, constraints, their associated jacobian matrices and dimensions of the problem.
 
 Although the package enables one to create linear unconstrained least squares, it is recommended to use it to solve nonlinear least squares with general constraints.
 
@@ -39,17 +43,17 @@ The three following positional arguments are mandatory to create a model:
 * `nb_parameters` : number of variables
 * `nb_residuals` : number of residuals
 
-The following keywords arguments are optionnal and deal with constraints and jacobian matrices computations. If the jacobian matrices functions are not provided, they are computed numerically by forward differences within `Enlsip`.
+The following keywords arguments are optional and deal with constraints and Jacobian matrices. If the Jacobian matrices functions are not provided, they are computed numerically by forward differences within the `Enlsip` solver.
 
  Argument             | Details
 :---------------------|:----------------------------------------------
  `starting_point`     | initial solution (can be an infeasbile point)
- `jacobian_residuals` | function computing the jacobian matrix of the residuals
+ `jacobian_residuals` | function computing the Jacobian matrix of the residuals
  `eq_constraints`     | function computing the equality constraints
- `jacobian_eqcons`    | function computing the jacobian matrix of the equality constraints
+ `jacobian_eqcons`    | function computing the Jacobian matrix of the equality constraints
  `nb_eqcons`          | number of equality constraints
  `ineq_constraints`   | function computing the inequality constraints
- `jacobian_ineqcons`  | function computing the jacobian matrix of the inequality constraints
[…]
