### 2024-02-27 #6226 @pierre-borie
@jbytecode So far I pushed corrections to the [documentation](https://uncertainlab.github.io/Enlsip.jl/dev/). I also added formal corrections to the paper but I want to include an industrial application of our package before submitting another draft. I am currently working on this and still need to fix what we are allowed to show or not wiht the industrial partner in question.
