### 2024-02-27 UncertainLab/Enlsip.jl#2 @pierre-borie
I modified the [documentation](https://uncertainlab.github.io/Enlsip.jl/dev/) by adding a description of the method and correcting typos. I also changed the `status` docstring eventhough it is not visible in the documentation page yet.
I am currently working with Hydro-Québec to provide an industrial use of our package.
