### 2024-02-27 UncertainLab/Enlsip.jl#3 @pierre-borie
Thank you for your feedback and sorry for taking this much time to respond.
I agree on your suggestion to clarify the theoritical guarantees but shouldn't I redirect to the [documentation](https://uncertainlab.github.io/Enlsip.jl/dev/) for a description of the method? I might be wrong about it but I wanted to keep the optimization aspects of the paper brief and just state in what category the method belongs to.
Also thanks for the topic comparing least squares solvers packages!
