### 2024-02-27 UncertainLab/Enlsip.jl#3 @tmigot
It is true, though but it is possible to give a short description of the algorithm in 2-3 sentences.
