### 2024-02-28 f0481c6
DisplayedInfo struc added + begin float type modification
