+++ README.md
@@ -1,9 +1,6 @@
 # Enlsip.jl
 
-
 [![](https://img.shields.io/badge/docs-dev-blue.svg)](https://uncertainlab.github.io/Enlsip.jl/dev/)
-
-
 
 Package `Enlsip.jl` is the Julia version of an optimization library originally written in Fortran77 and designed to solve nonlinear least squares problems under nonlinear constraints.
  The optimization method implemented in Enlsip was conceived by two swedish authors, Per Lindstrom and Per Ake Wedin from the Institute of Informatation processing of the University of Umea in Sweden.
@@ -71,7 +68,6 @@ An object of type `CnlsModel` can be created using a constructor, whose argument
 ### Solving a model
 
 Then, once your model is instantiated, you can call the `solve!` function to solve your problem.
-
 
 ### Example with problem 65 from  Hock Schittkowski collection
 
