### 2024-02-29 60f4e20
compatilibility extended to float subtypes of AbstractFloat
