+++ test/chained_rosenbrock.jl
@@ -52,7 +52,7 @@
     
     x0 = [(mod(i,2) == 1 ? -1.2 : 1.0) for i=1:n]
 
-    Crmodel = CnlsModel(r,n,m ;starting_point=x0, jacobian_residuals=jac_res, eq_constraints=c, jacobian_eqcons=jac_cons, nb_eqcons=nb_eq)
+    Crmodel = CnlsModel(r,n,m; starting_point=x0, jacobian_residuals=jac_res, eq_constraints=c, jacobian_eqcons=jac_cons, nb_eqcons=nb_eq)
 
     CR_sol = solve!(Crmodel)
 
