+++ src/enlsip_functions.jl
@@ -546,7 +546,7 @@ function minmax_lagrangian_mult(
     q,t = working_set.q, working_set.t
     scaling = active_C.scaling
     diag_scale = active_C.diag_scale
-    sq_rel = sqrt(eps(Float64))
+    sq_rel = sqrt(eps(eltype(λ)))
     λ_abs_max = 0.0
     sigmin = 1e6
 
@@ -575,17 +575,16 @@ function check_constraint_deletion(
     q::Int,
     A::Matrix{<:AbstractFloat},
     λ::Vector{<:AbstractFloat},
-    ∇fx::Vector{<:AbstractFloat},
     scaling::Bool,
     diag_scale::Vector{<:AbstractFloat},
     grad_res::AbstractFloat)
 
       
-    (t, n) = size(A)
+    t = size(A, 1)
     δ = 10.0
     τ = 0.5
     λ_max = (isempty(λ) ? 1.0 : maximum(map(t -> abs(t), λ)))
-    sq_rel = sqrt(eps(Float64)) * λ_max
+    sq_rel = sqrt(eps(eltype(λ))) * λ_max
     s = 0
     
     if t > q
@@ -613,7 +612,7 @@ function evaluate_violated_constraints(
     index_α_upp::Int)
 
     # Data
-    ε = sqrt(eps(Float64))
+    ε = sqrt(eps(eltype(cx)))
     δ = 0.1
     added = false
     if W.l > W.t
@@ -720,7 +719,7 @@ function update_working_set(
     F_A = qr(C.A', ColumnNorm())
     
     first_lagrange_mult_estimate!(C.A, λ, ∇fx, C.cx, C.scaling, C.diag_scale, F_A, iter_k, ε_rank)
-    s = check_constraint_deletion(W.q, C.A, λ, ∇fx, C.scaling, C.diag_scale, iter_k.grad_res)
+    s = check_constraint_deletion(W.q, C.A, λ, C.scaling, C.diag_scale, iter_k.grad_res)
     (m, n) = size(J)
     # Constraint number s is deleted from the current working set
     if s != 0
@@ -733,7 +732,7 @@ function update_working_set(
         deleteat!(λ, s)
         deleteat!(C.cx, s)
         deleteat!(C.diag_scale, s)
-        delete_constraint!(W, s)
+        remove_constraint!(W, s)
         iter_k.del = true
         iter_k.index_del = index_s
         C.A = C.A[setdiff(1:end, s), :]
@@ -764,13 +763,13 @@ function update_working_set(
 
             if !(W.t != rankA || iter_k.rankJ2 != min(m, n - rankA))
                 second_lagrange_mult_estimate!(J, F_A, λ, rx, p_gn, W.t, C.scaling, C.diag_scale)
-                s2 = check_constraint_deletion(W.q, C.A, λ, ∇fx, C.scaling, C.diag_scale, 0.0)
+                s2 = check_constraint_deletion(W.q, C.A, λ, C.scaling, C.diag_scale, 0.0)
                 if s2 != 0
                     index_s2 = W.active[s2]
                     deleteat!(λ, s2)
                     deleteat!(C.diag_scale, s2)
                     C.cx = C.cx[setdiff(1:end, s2)]
-                    delete_constraint!(W, s2)
+                    remove_constraint!(W, s2)
                     iter_k.del = true
                     iter_k.index_del = index_s2
                     C.A = C.A[setdiff(1:end, s2), :]
@@ -792,13 +791,13 @@ function update_working_set(
 
[…]
