### 2024-02-29 994f571
iteration detail can be printed after solving the problem
