+++ src/enlsip_functions.jl
@@ -2499,172 +2499,58 @@ function check_termination_criteria(
     return exit_code
 end
 
-# OUTPUT
-# Print the useful informations at the end of current iteration
+#=
+Functions to print details about execution of the algorithm
+=#
 
-function output!(
-    io::IOStream,
-    iter::Iteration,
-    W::WorkingSet,
-    nb_iter::Int,
-    rx_sum::AbstractFloat,
-    cx_sum::AbstractFloat)
+function print_header(m::Int, n::Int, q::Int, l::Int, scaling::Bool, io::IO=stdout)
+    print(io,"\n\n")
+        println(io, '*'^64)
+        println(io, "*",' '^62,"*")
 
-    if norm(W.active, Inf) > 0
-        s_act = "("
-        # Pour ne pas afficher trop d'indices
-        for i = 1:min(5,W.t)
-            s_act = (i < W.t ? string(s_act, W.active[i], ",") : string(s_act, W.active[i], ")"))
-        end
-    else
-        s_act = " -"
-    end
-    speed = (nb_iter == 0 ? 0.0 : iter.speed) # iter.β / β_prev
-    @printf(io, "  %2d  %e  %.2e  %9.2e   %.3e %3d   %3d   %.2e    %.2e     %.2e    %s\n", nb_iter, rx_sum, cx_sum, iter.progress, norm(iter.p), iter.dimA, iter.dimJ2, iter.α, speed, maximum(iter.w), s_act)
+        println(io, "*                          Enlsip 0.9.1                        *")
+        println(io, "*",' '^62,"*")
+        println(io, "* Julia version of a Fortran77 solver conceived and developed  *")
+        println(io, "* by Per Lindstrom and Per Ake Wedin from the Institute        *")
+        println(io, "* of Information processing, University of Umea, Sweden.       *")
+        println(io, "*",' '^62,"*")
+        println(io, '*'^64)
+
+        println(io, "\n\nNumber of parameters                 : ", @sprintf("%5i", n))
+        println(io, "Number of residuals                  : ", @sprintf("%5i", m))
+        println(io, "Number of equality constraints       : ", @sprintf("%5i", q))
+        println(io, "Number of inequality constraints     : ", @sprintf("%5i", (l - q)))
+        println(io, "Constraints internal scaling         : $scaling\n")
+        println(io, "\nIteration steps information\n")
+        println(io, "iter     objective    ||active_constraints||²    ||p||        α     reduction")
 end
 
-function final_output!(
-    io::IOStream,
-    iter::Iteration,
-    W::WorkingSet,
-    exit_code::Int,
-    nb_iter::Int)
-
-
-    
-    @printf(io, "\nExit code : %d\nNumber of iterations : %d \n\n", exit_code, nb_iter)
-    print(io, "Terminated at point :")
-    (t -> @printf(io, " %e ", t)).(iter.x)
-    print(io, "\n\nActive constraints :")
-    (i -> @printf(io, " %d ", i)).(W.active[1:W.t])
-    println(io, "\nConstraint values : ")
-    (t -> @printf(io, " %.2e ", t)).(iter.cx)
-    println(io, "\nPenalty constants :")
-    (t -> @printf(io, " %.2e ", t)).(iter.w)
-
-    @printf(io, "\n\nSquare sum of residuals = %e\n\n", dot(iter.rx, iter.rx))
+function print_iter(k::Int, iter_data::DisplayedInfo; io::IO=stdout)
+    @printf(io, "%4d   %.7e        %.2e           %.2e   %.2e   %.2e\n", k, iter_data.objective, iter_data.sqr_nrm_act_cons, 
[…]
