### 2024-02-29 aeb156a
documentation for the print_cnls_model function
