+++ test/chained_rosenbrock.jl
@@ -53,8 +53,9 @@
     x0 = [(mod(i,2) == 1 ? -1.2 : 1.0) for i=1:n]
 
     Crmodel = CnlsModel(r,n,m; starting_point=x0, jacobian_residuals=jac_res, eq_constraints=c, jacobian_eqcons=jac_cons, nb_eqcons=nb_eq)
-
+    print_cnls_model(Crmodel)
     solve!(Crmodel)
+    print_cnls_model(Crmodel)
 
 
     @test size(x0,1) == Crmodel.nb_parameters
@@ -66,7 +67,5 @@
     @test nb_constraints == total_nb_constraints(Crmodel)
     @test status(Crmodel) in values(dict_status_codes)
     @test typeof(solution(Crmodel)) <: Vector && size(solution(Crmodel),1) == n
-    @test typeof(objective_value(Crmodel)) <: Number && isfinite(objective_value(Crmodel))
-  
-    
+    @test typeof(objective_value(Crmodel)) <: Number && isfinite(objective_value(Crmodel))  
 end
