+++ test/HS65.jl
@@ -18,8 +18,9 @@
 
 
     hs65_model = Enlsip.CnlsModel(r, n, m ;jacobian_residuals=jac_r, starting_point=x0, ineq_constraints = c, jacobian_ineqcons=jac_c, nb_ineqcons = 1, x_low=x_l, x_upp=x_u)
-    solve!(hs65_model, silent=false)
-
+    print_cnls_model(hs65_model)
+    solve!(hs65_model)
+    print_cnls_model(hs65_model)
 
     @test size(x0,1) == hs65_model.nb_parameters
     @test r(x0) ≈ hs65_model.residuals(x0)
@@ -29,6 +30,4 @@
     @test status(hs65_model) in values(dict_status_codes)
     @test typeof(solution(hs65_model)) <: Vector && size(solution(hs65_model),1) == n
     @test typeof(objective_value(hs65_model)) <: Number && isfinite(objective_value(hs65_model))
-
-
 end
