### 2024-03-04 f34762f
Corrected methods definition for AbtrsctFloat subtypes
