+++ src/enlsip_functions.jl
@@ -14,7 +14,7 @@ Parameters :
     the number of rows of matrix `T`.
 =#
 
-function pseudo_rank(diag_T::Vector{<:AbstractFloat}, ε_rank::AbstractFloat)
+function pseudo_rank(diag_T::Vector{T}, ε_rank::T) where {T}
 
     if isempty(diag_T) || abs(diag_T[1]) < ε_rank
         pseudo_rank = 0
@@ -32,13 +32,13 @@ end
 
 
 function new_point!(
-    x::Vector,
+    x::Vector{T},
     r::ResidualsFunction,
     c::ConstraintsFunction,
-    rx::Vector,
-    cx::Vector,
-    J::Matrix,
-    A::Matrix)
+    rx::Vector{T},
+    cx::Vector{T},
+    J::Matrix{T},
+    A::Matrix{T}) where {T}
 
     # Evaluate residuals and associated jacobian matrix
     res_eval!(r,x,rx)
@@ -125,7 +125,7 @@ function sub_search_direction(
     rankA::Int,
     dimA::Int,
     dimJ2::Int,
-    code::Int) where {T<:AbstractFloat}
+    code::Int) where {T}
 
     # Solving without stabilization 
     if code == 1
@@ -134,18 +134,18 @@ function sub_search_direction(
         d_temp = -J1 * p1 - rx
         d = F_J2.Q' * d_temp
         δp2 = UpperTriangular(F_J2.R[1:dimJ2, 1:dimJ2]) \ d[1:dimJ2]
-        p2 = [δp2; zeros(n - t - dimJ2)][invperm(F_J2.p)]
+        p2 = [δp2; zeros(T, n - t - dimJ2)][invperm(F_J2.p)]
 
     # Solving with stabilization
     elseif code == -1
         b_buff = -cx[F_A.p]
         b = F_L11.Q' * b_buff
         δp1 = UpperTriangular(F_L11.R[1:dimA, 1:dimA]) \ b[1:dimA]
-        p1 = F_L11.P[1:rankA, 1:rankA] * [δp1; zeros(rankA - dimA)]
+        p1 = ([δp1; zeros(t - dimA)][invperm(F_L11.p)])[1:rankA]
         d_temp = -J1 * p1 - rx
         d = F_J2.Q' * d_temp
         δp2 = UpperTriangular(F_J2.R[1:dimJ2, 1:dimJ2]) \ d[1:dimJ2]
-        p2 = [δp2; zeros(n - rankA - dimJ2)][invperm(F_J2.p)]
+        p2 = [δp2; zeros(T, n - rankA - dimJ2)][invperm(F_J2.p)]
     end
 
     p = F_A.Q * [p1; p2]
@@ -204,18 +204,18 @@ If `rankA = t`, the first system is solved, otherwise, the second one is solved.
 * `F_J2` : QR decomposition of Matrix `J2` defined in [`sub_search_direction`](@ref)
 =#
 function gn_search_direction(
-    J::Matrix{<:AbstractFloat},
-    rx::Vector{<:AbstractFloat},
-    cx::Vector{<:AbstractFloat},
+    J::Matrix{T},
+    rx::Vector{T},
+    cx::Vector{T},
     F_A::Factorization,
     F_L11::Factorization,
     rankA::Int,
     t::Int,
-    ε_rank::AbstractFloat,
-    current_iter::Iteration)
+    ε_rank::T,
+    current_iter::Iteration) where {T}
 
     code = (rankA == t ? 1 : -1)
-    (m, n) = size(J)
+    n = size(J,2)
     JQ1 = J * F_A.Q
     J1, J2 = JQ1[:, 1:rankA], JQ1[:, rankA+1:end]
[…]
