### 2024-03-04 dc0c226
added tests on the WorkingSet structure
