+++ test/runtests.jl
@@ -1,6 +1,7 @@
 using Enlsip
 using Test
 
- 
-include("chained_rosenbrock.jl")
+include("internal_tests.jl") 
 include("HS65.jl")
+include("chained_rosenbrock.jl")
+
