### 2024-03-05 80c9837
add AD option for residuals + constraints evaluation function for a solved problem
