+++ test/runtests.jl
@@ -1,7 +1,8 @@
 using Enlsip
 using Test
 
-include("internal_tests.jl") 
-include("HS65.jl")
-include("chained_rosenbrock.jl")
+# include("internal_tests.jl") 
+# include("HS65.jl")
+include("osborne2.jl")
+# include("chained_rosenbrock.jl")
 
