+++ test/chained_rosenbrock.jl
@@ -8,7 +8,7 @@
     function r(x::Vector)
         n = length(x)
         m = 2(n-1)
-        rx = Vector(undef,m)
+        rx = Vector{eltype(x)}(undef,m)
         rx[1:n-1] = [10(x[i]^2 - x[i+1]) for i=1:n-1]
         rx[n:m] = [x[k-n+1] - 1 for k=n:m]
         return rx
@@ -52,11 +52,10 @@
     
     x0 = [(mod(i,2) == 1 ? -1.2 : 1.0) for i=1:n]
 
-    Crmodel = CnlsModel(r,n,m; starting_point=x0, jacobian_residuals=jac_res, eq_constraints=c, jacobian_eqcons=jac_cons, nb_eqcons=nb_eq)
+    Crmodel = CnlsModel(r,n,m; starting_point=x0, jacobian_residuals = jac_res, eq_constraints=c, jacobian_eqcons=jac_cons, nb_eqcons=nb_eq)
     print_cnls_model(Crmodel)
     solve!(Crmodel)
     print_cnls_model(Crmodel)
-
 
     @test size(x0,1) == Crmodel.nb_parameters
     @test r(x0) ≈ Crmodel.residuals(x0)
