### 2024-03-05 6e6b932
Update Project.toml (Polynomials and ForwardDiff)
