### 2024-03-06 1e9bc09
tests on non exported function
