+++ src/cnls_model.jl
@@ -1,5 +1,6 @@
 export AbstractCnlsModel, CnlsModel
-export total_nb_constraints, status, solution, objective_value, dict_status_codes
+export constraints_values, equality_constraints_values, inequality_constraints_values, bounds_constraints_values, total_nb_constraints, nb_equality_constraints, nb_inequality_constraints, nb_lower_bounds, nb_upper_bounds
+export status, solution, objective_value, dict_status_codes
 
 #= Structures where the first two fields are the functions evaluating residuals or constraints and the associated jacobian matrix
 =#
@@ -226,6 +227,14 @@ If no convergence, this value is computed at the last solution obtained.
 """
 objective_value(model::CnlsModel) = model.obj_value
 
+nb_equality_constraints(model::CnlsModel) = model.nb_eqcons
+
+nb_inequality_constraints(model::CnlsModel) = model.nb_ineqcons
+
+nb_lower_bounds(model::CnlsModel) = count(isfinite, model.x_low)
+
+nb_upper_bounds(model::CnlsModel) = count(isfinite, model.x_upp)
+
 """
     total_nb_constraints(model)
 
@@ -237,13 +246,13 @@ total_nb_constraints(model::CnlsModel) = model.nb_eqcons + model.nb_ineqcons + c
 
 
 """
-    equality_constrains_values(model)
-    
+    equality_constraints_values(model)
+
 Returns the vector of equality constraints values at the solution of `model` (if they are any).
 """
-function equality_constrains_values(model::CnlsModel)
+function equality_constraints_values(model::CnlsModel)
     sol = solution(model)
-    hx = eltype{sol}[]
+    hx = Vector{eltype(sol)}([])
     if model.eq_constraints !== nothing
         hx = model.eq_constraints(sol)
     end
@@ -257,9 +266,9 @@ Returns the vector of inequality constraints values at the solution of `model` (
 """
 function inequality_constraints_values(model::CnlsModel)
     sol = solution(model)
-    gx = eltype{sol}[]
-    if model.eq_constraints !== nothing
-        gx = model.eq_constraints(sol)
+    gx = Vector{eltype(sol)}([])
+    if model.ineq_constraints !== nothing
+        gx = model.ineq_constraints(sol)
     end
     return gx
 end
@@ -267,17 +276,24 @@ end
 """
     bounds_constraints_values(model)
 
-Returns the vector of box constraints values at the solution of `model` (if they are any).   
+Returns the vector of box constraints values at the solution `xₛ` of `model` (if they are any).   
+
+If `xₗ` and `xᵤ` are respectively the vectors of lower and upper bounds, it will return `[xₛ-xₗ; xᵤ-xₛ]`.
 """
 function bounds_constraints_values(model::CnlsModel)
     sol = solution(model)
-    return vcat(filter(isfinite, sol-model.x_low), filter(isfinite, model.x_upp-sol))
+    return vcat(sol-model.x_low, model.x_upp-sol)
 end
 
 """
     constraints_values(model)
 
-Returns the vector of values of all the constraints in `model` at the solution. 
+Computes values of all the constraints in `model` at the solution. 
+
+The vector returned is the concatenation of equalities, inequalities and box constraints (in that order).
+
[…]
