+++ test/runtests.jl
@@ -1,8 +1,10 @@
 using Enlsip
 using Test
 
-# include("internal_tests.jl") 
-# include("HS65.jl")
-include("osborne2.jl")
-# include("chained_rosenbrock.jl")
+
+include("internal/working_set.jl")
+include("internal/constraints.jl") 
+include("problems/HS65.jl")
+include("problems/osborne2.jl")
+include("problems/chained_rosenbrock.jl")
 
