+++ versions.js
@@ -1,5 +1,7 @@
 var DOC_VERSIONS = [
+  "stable",
+  "v0.9",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "dev";
-var DOCUMENTER_STABLE = "dev";
+var DOCUMENTER_NEWEST = "v0.9.2";
+var DOCUMENTER_STABLE = "stable";
