### 2024-03-09 69be47f
correct typos in enlsip.functions.jl
