+++ src/enlsip_functions.jl
@@ -368,7 +368,7 @@ function newton_search_direction(
         b = -active_cx[F_A.p]
         p1 = LowerTriangular(F_A.R') \ b
     elseif t > rankA
-        b = -F_L11.Q' * active_cx[F_A.p]
+        b = F_L11.Q' * (-active_cx[F_A.p])
         δp1 = UpperTriangular(F_L11.R[1:rankA, 1:rankA]) \ b[1:rankA]
         p1 = F_L11.P[1:rankA, 1:rankA] * δp1
     end
@@ -412,7 +412,7 @@ function newton_search_direction(
         p2 = chol_W22.U \ y
         p = F_A.Q * [p1; p2]
     else
-        p = zeros(n)
+        p = zeros(T, n)
         error = true
     end
     return p, error
@@ -461,7 +461,7 @@ function first_lagrange_mult_estimate!(
     cx::Vector{T},
     scaling_done::Bool,
     diag_scale::Vector{T},
-    F::Factorization{T},
+    F::Factorization,
     iter::Iteration,
     ε_rank::T) where {T} 
 
@@ -536,7 +536,6 @@ function minmax_lagrangian_mult(
     active_C::Constraint) where {T}
     
     # Data
-
     q,t = working_set.q, working_set.t
     scaling = active_C.scaling
     diag_scale = active_C.diag_scale
@@ -572,8 +571,7 @@ function check_constraint_deletion(
     scaling::Bool,
     diag_scale::Vector{T},
     grad_res::T) where {T}
-
-      
+  
     t = size(A, 1)
     δ = 10.0
     τ = 0.5
@@ -616,50 +614,12 @@ function evaluate_violated_constraints(
             if cx[k] < ε || (k == index_α_upp && cx[k] < δ)
                 add_constraint!(W, i)
                 added = true
-                if added
-                end
             else
                 i += 1
             end
         end
     end
     return added
-end
-
-# Updates QR factorisation of A^T by appyling Givens rotations
-
-function update_QR_A(A::Matrix{T}) where {T}
-    (t,n) = size(A)
-    F_A = qr(A', ColumnNorm()) 
-    Q1 = F_A.Q*Matrix(I,n,n)
-    L11, P1 = Matrix(F_A.R'), F_A.P
-    return P1, L11, Q1
-end
-
-function update_QR_A(
-    Q::Matrix{T},
-    R::Matrix{T},
-    p::Vector{Int},
-    s::Int,
-    t::Int) where {T}
-
-    # Update permutation vector, form permutation matrix, delete j-th column of matrix R 
-    j = p[s]
-    setdiff!(p, j)
-    p[:] = [(e > j ? e - 1 : e) for e in p] 
-    P = (1.0*Matrix(I, t, t))[:, p]
-
-    R_temp = R[:, setdiff(1:end, j)]
-
-    # Apply Givens rotations to transform R to Upper triangular form
-
-    for i = j:t
-        G, r = givens(-R_temp[i, i], -R_temp[i+1, i], i, i + 1)
-        lmul!(G, R_temp)
-        rmul!(Q, G')
-    end
-
-    return P, Matrix(transpose(R_temp[1:t, 1:t])), Q, p
 end
 
 
@@ -1259,7 +1219,7 @@ function search_direction_analys(
     elseif method_code == -1
         JQ1 = J * F_A.Q
         J1 = JQ1[:, 1:rankA]
-        b = -F_L11.Q' * active_cx[F_A.p]
+        b = F_L11.Q' * (-active_cx[F_A.p])
         dimA, dimJ2 = choose_subspace_dimensions(rx_sum, rx, active_cx_sum, J1, working_set.t, rankJ2, rankA, b, F_L11, F_J2, previous_iter, restart)
[…]
