### 2024-03-22 9d3d05c
Corrected bug in method call for hessian matrix computation
