+++ test/runtests.jl
@@ -7,4 +7,5 @@ include("internal/constraints.jl")
 include("problems/HS65.jl")
 include("problems/osborne2.jl")
 include("problems/chained_rosenbrock.jl")
+include("problems/chained_wood.jl")
 
