### 2024-03-22 #6226 @pierre-borie
@jbytecode For the past weeks I have encountered difficulties in the process of having access to an industrial application that can be described in the paper. Thoses issues have been fixed so I should be able to submit a new draft next week.

Concerning the package, In this commit 855702b4a94afd9c6755bb6f41fc9d58e094fa9b I added and corrected some functionalities following advices from the reviewers. [Documentation](https://uncertainlab.github.io/Enlsip.jl/stable/) has been modified accordingly to thoses changes.
