### 2024-03-28 310b089
zeros vectors initialized with abstract type
