+++ src/enlsip_functions.jl
@@ -141,7 +141,7 @@ function sub_search_direction(
         b_buff = -cx[F_A.p]
         b = F_L11.Q' * b_buff
         δp1 = UpperTriangular(F_L11.R[1:dimA, 1:dimA]) \ b[1:dimA]
-        p1 = ([δp1; zeros(t - dimA)][invperm(F_L11.p)])[1:rankA]
+        p1 = ([δp1; zeros(T,t - dimA)][invperm(F_L11.p)])[1:rankA]
         d_temp = -J1 * p1 - rx
         d = F_J2.Q' * d_temp
         δp2 = UpperTriangular(F_J2.R[1:dimJ2, 1:dimJ2]) \ d[1:dimJ2]
@@ -970,7 +970,7 @@ function check_gn_direction(
         to_reduce = false
         if W.q < W.t
             sqr_ε = sqrt(eps(eltype(λ)))
-            rows = zeros(W.t - W.q)
+            rows = zeros(T, W.t - W.q)
             for i = W.q+1:W.t
                 rows[i-W.q] = (scaling ? 1.0 / diag_scale[i] : diag_scale[i])
             end
@@ -1029,7 +1029,7 @@ function determine_solving_dim(
     mindim = 1
 
     if rankR > 0
-        l_estim_sd, l_estim_righthand = zeros(rankR), zeros(rankR)
+        l_estim_sd, l_estim_righthand = zeros(T,rankR), zeros(T,rankR)
         l_estim_sd[1] = abs(y[1])
         l_estim_righthand[1] = abs(y[1] / R[1, 1])
 
@@ -1113,7 +1113,8 @@ function choose_subspace_dimensions(
         d = -rx
 
     elseif rankA > 0
-        previous_dimA = abs(previous_iter.rankA) + t - previous_iter.t
+        # previous_dimA = abs(previous_iter.rankA) + t - previous_iter.t
+        previous_dimA = abs(previous_iter.dimA) + t - previous_iter.t
         nrm_b_asprev = norm(b[1:previous_dimA])
         nrm_b = norm(b)
         constraint_progress = dot(previous_iter.cx, previous_iter.cx) - active_cx_sum
@@ -1126,7 +1127,7 @@ function choose_subspace_dimensions(
         # Forms right hand side d = r(x)+J1*p1
 
         δp1 = UpperTriangular(F_L11.R[1:dimA, 1:dimA]) \ b[1:dimA]
-        p1 = F_L11.P[1:rankA, 1:rankA] * [δp1; zeros(rankA - dimA)]
+        p1 = F_L11.P[1:rankA, 1:rankA] * [δp1; zeros(T,rankA - dimA)]
         d = -(rx + J1 * p1)
     end
 
@@ -1134,7 +1135,8 @@ function choose_subspace_dimensions(
         d = F_J2.Q' * d
     end
 
-    previous_dimJ2 = abs(previous_iter.rankJ2) + previous_iter.t - t
+    # previous_dimJ2 = abs(previous_iter.rankJ2) + previous_iter.t - t
+    previous_dimJ2 = abs(previous_iter.dimJ2) + previous_iter.t -t
     nrm_d_asprev = norm(d[1:previous_dimJ2])
     nrm_d = norm(d)
     residual_progress = dot(previous_iter.rx, previous_iter.rx) - rx_sum
@@ -1431,7 +1433,7 @@ function euclidean_norm_weight_update(
         elseif (ztw < μ) && (dimA < t)
 
             # Form vector e and scalar τ (\tau)
-            e = zeros(t)
+            e = zeros(T,t)
             ctrl, nb_pos, τ = 2, 0, μ
             for i = 1:t
                 k = active[i]
@@ -1643,7 +1645,7 @@ function coefficients_linesearch!(v0::Vector{T},
[…]
