+++ paper/paper.md
@@ -29,58 +29,89 @@ bibliography: paper.bib
 
 # Summary
 
-Easy Nonlinear Least Squares Inequality Program (``ENLSIP``[^1]) is the name of an optimization algorithm and an open-source Fortran77 library developed and released in 1988. It implements a nonlinear least squares under nonlinear constraints solver.
+ [Enlsip.jl](https://github.com/UncertainLab/Enlsip.jl) is a Julia  [@Julia:2017] package that implements a solver for nonlinear least squares with nonlinear constraints.
 
 This type of problems is mathematically formulated as:
-\begin{subequations}\label{eq:cnlls}
-         \quad \begin{align}  
+\begin{equation}\label{eq:cnlls}
+         \quad \begin{aligned}  
                         \quad	\min_{x \in \mathbb{R}^n}        \quad&  \dfrac{1}{2} \sum_{i=1}^{m} r_i(x)^2  \\
-                        \text{s.t.}      \quad & c_i(x)=0, \quad 1\leq i \leq q \\
-                        & c_i(x) \geq 0, \quad q+1 \leq i \leq \ell,
-        \end{align}
-\end{subequations}
+                        \text{s.t.}      \quad & c_i(x)=0, \quad i \in \mathcal{E}\\
+                        & c_i(x) \geq 0, \quad i\in\mathcal{I},
+        \end{aligned}
+\end{equation}
 
-where the multi-functions $r_i$, often denoted as the residuals, and constraints $c_i$ are two-times differentiable functions. Integers $n,m,q$ and $l$ are the dimensions of the problem.
+where the $m$ functions $r_i$, often denoted as the residuals, and constraints $c_i$ are two-times differentiable functions.
 
-The ``ENLISP`` solver incorporates a Gauss-Newton type method developed by @lindstromwedin1988. This method also uses an active set strategy to handle the constraints [see @nocedalwright:2006, chapter 16].
+This package is the Julia version of ``ENLSIP`` (Easy Nonlinear Least Squares Inequality Program), an open-source Fortran77 library developed by @lindstromwedin1988[^1].
 
 [^1]: The source code is available at [https://plato.asu.edu/sub/nonlsq.html](https://plato.asu.edu/sub/nonlsq.html)
 
 # Statement of need
 
-The ``ENLSIP`` Fortran77 library has been successfully used for decades by Hydro-Québec, the main electricity supplier for the province of Quebec in Canada, to calibrate its short-term electricity demand forecast models, which are coded in Fortran90. Since Hydro-Québec is transitioning from Fortran77 to Julia [@Julia:2017] and because its systems are used in a highly critical context, the primary goal of this transition is to ensure that the replacing Julia version reproduces the results given by the original Fortran77 version. The conversion of the above-mentioned ``ENLSIP`` library to Julia is a crucial part of this process.
+The ``ENLSIP`` Fortran77 library has been successfully used since the early 2000s by Hydro-Québec, the main electricity supplier for the province of Quebec in Canada, to calibrate its short-term electricity demand forecast models [@grenieretal:2006], which are coded in Fortran90. Since Hydro-Québec is transitioning from Fortran77 to Julia and because its systems are used in a highly critical context, the primary goal of this transition is to ensure that the replacing Julia version reproduces the results given by the original Fortran77 version. The conversion of the above-mentioned ``ENLSIP`` library to Julia is a crucial part of this process.
 
-Nonlinear least squares arise in a variety of model calibration scenarios. Formulation \eqref{eq:cnlls} is particularly relevant in contexts where additional constraints, such as those related to physical models, need to be satisfied. This is due to the high-risk nature of Hydro-Québec's forecasting operations.
[…]
