+++ paper/paper.bib
@@ -1,3 +1,12 @@
+@inproceedings{grenieretal:2006,
+  author={M. Grenier and J. Gagnon and C. Mercier and J. Richard},
+  booktitle={2006 IEEE Power Engineering Society General Meeting},
+  title={Short-term load forecasting at {Hydro-Québec TransÉnergie}},
+  year={2006},
+  numpages={5},
+  address = {Montreal, QC, Canada},
+  doi={10.1109/PES.2006.1709029}}
+
 @article{lindstromwedin1984,
   author  = {P. Lindstr{\"o}m and P.{\AA}. Wedin},
   journal = {Mathematical Programming},
@@ -108,6 +117,14 @@ journal={Technical report 767}
   doi       = {10.1137/1.9781611971200},
   series    = {Classics in Applied Mathematics},
 }
+
+@book{gillmurraywright:1985,
+ 	author    = {P.E. Gill and W. Murray and M.H. Wright},
+ 	publisher = {Academic Press},
+ 	address   = {San Diego, CA, USA}, 
+ 	title     = {Practical Optimization},
+ 	year      = {1981},
+ 	doi       = {10.1007/978-0-387-40065-5}}
 
 @book{hockschittkowski:1980,
   author    = {W. Hock and K. Schittkowski},
