+++ paper/paper.md
@@ -35,12 +35,12 @@ This type of problems is mathematically formulated as:
 \begin{equation}\label{eq:cnlls}
          \quad \begin{aligned}  
                         \quad	\min_{x \in \mathbb{R}^n}        \quad&  \dfrac{1}{2} \sum_{i=1}^{m} r_i(x)^2  \\
-                        \text{s.t.}      \quad & c_i(x)=0, \quad i \in \mathcal{E}\\
-                        & c_i(x) \geq 0, \quad i\in\mathcal{I},
+                        \text{s.t.}      \quad & c_i(x)=0, \quad i =1,\ldots,q\\
+                        & c_i(x) \geq 0, \quad i=q+1,\ldots,\ell,
         \end{aligned}
 \end{equation}
 
-where the $m$ functions $r_i$, often denoted as the residuals, and constraints $c_i$ are two-times differentiable functions.
+where the functions $(r_1,\ldots,r_m)$, often denoted as the residuals, and constraints $(c_1,\ldots,c_\ell)$ are two-times differentiable.
 
 This package is the Julia version of ``ENLSIP`` (Easy Nonlinear Least Squares Inequality Program), an open-source Fortran77 library developed by @lindstromwedin1988[^1].
 
@@ -50,7 +50,7 @@ This package is the Julia version of ``ENLSIP`` (Easy Nonlinear Least Squares In
 
 The ``ENLSIP`` Fortran77 library has been successfully used since the early 2000s by Hydro-Québec, the main electricity supplier for the province of Quebec in Canada, to calibrate its short-term electricity demand forecast models [@grenieretal:2006], which are coded in Fortran90. Since Hydro-Québec is transitioning from Fortran77 to Julia and because its systems are used in a highly critical context, the primary goal of this transition is to ensure that the replacing Julia version reproduces the results given by the original Fortran77 version. The conversion of the above-mentioned ``ENLSIP`` library to Julia is a crucial part of this process.
 
-Nonlinear least squares arise in a variety of model calibration scenarios. Formulation \eqref{eq:cnlls} is particularly relevant in contexts where additional constraints, such as those related to physical models, need to be satisfied. This is due to the high-risk nature of Hydro-Québec's forecasting operations. Moreover, the specific structure of this type of problems can be exploited to design algorithms more efficient than solvers for general nonlinear programming.
+Nonlinear least squares arise in a variety of model calibration scenarios. Formulation \eqref{eq:cnlls} is particularly relevant in contexts where additional constraints, such as those related to physical models, need to be satisfied. This is due to the high-risk nature of Hydro-Québec's forecasting operations. Moreover, the specific structure of the objective function in \eqref{eq:cnlls} can be exploited to design algorithms more efficient than solvers for general nonlinear programming.
 
 Comparison of results and performance on operational Hydro-Québec optimization problems have been conducted using a Julia-Fortran interface and they have shown very good concordance results. We additionally compared numerical results on nonlinear programming test problems [@hockschittkowski:1980; @lucksanvlcek:1999] to ascertain whether the two versions could yield significantly disparate outcomes or distinct solutions. On the tested problems, we observed no differences in convergence behavior. Furthermore, the obtained solutions did not differ from a predetermined tolerance, the same one we previously employed to ensure the results of our Julia version were consistent with the requirements of Hydro-Québec. This has led us to consider that the current version of our implementation can be published as a registered Julia package.
 
@@ -58,7 +58,7 @@ Comparison of results and performance on operational Hydro-Québec optimization 
 
[…]
