### 2024-04-03 d30d9bb
Added example to documentation
