### 2024-04-03 #6226 @pierre-borie
@jbytecode  - It took a few more days than I thought it would but I pushed a revised version of the paper on the joss-paper branch.
