### 2024-04-03 UncertainLab/Enlsip.jl#3 @pierre-borie
Hi! I just pushed a revised version of the paper of the joss-paper branch. I added a brief description of the algorithm and the theoretical guarantees in addition to a paragraph on comparison tests between Julia and Fortran.
