### 2024-04-03 UncertainLab/Enlsip.jl#2 @pierre-borie
Hi! I pushed a revised version of the paper. A brief description of the method and its theoretical guarantees have been added. There is also a paragraph about an Hydro-Québec application.
