### 2024-04-10 UncertainLab/Enlsip.jl#2 @odunbar
Thanks @pierre-borie

I think I have ticked off everything that I can clearly see have been completed. I also understand that the HQ example is tough to use due to the proprietary parts.

- Please point me to the other items if I have missed them.
- I have a couple small additional items based on your corrections (see bottom of my issue above)
