### 2024-04-10 #6226 @odunbar
Update on my end, I had another sweep through my checklist:
@pierre-borie has addressed the majority of my points. I think the softening of claims in the the paper, the more comprehensive state of field, and the numerical example together give a more accurate link between what is claimed and evidenced.

There are a few typos and minor comments outstanding etc. but I do not expect these to take a long time.
