### 2024-04-15 44d0b79
doc: added chained rosenbrock table
