### 2024-04-15 UncertainLab/Enlsip.jl#2 @pierre-borie
@odunbar - Thanks for your positive feedback on the new version. From what I can see, there is just the "diagnostics" issue that has not been checked though the corresponding functionality has been added but it might be my fault. I wanted to mention this issue specifically but did wrong somewhere and tried to backtrack.

I corrected the paper accordingly to your last items and added a table for Chained Rosenbrock problem showing time for different dimensions on the dev version of the [documentation](https://uncertainlab.github.io/Enlsip.jl/dev/).

For now, I am waiting for the other reviewer's comments before submiting a revised version.
