+++ paper/paper.bib
@@ -77,6 +77,17 @@ pages = {348–368},
 numpages = {21}
 }
 
+@article{dolanmore:2002,
+author = {E.D. Dolan and J.J. Mor\'e},
+title = {Benchmarking optimization software with performance profiles},
+year = {2002},
+journal = {Mathematical Programming},
+volume = {Series A 91},
+number = {2},
+pages = {201-213},
+doi = {10.1007/s101070100263}}
+
+
 
 @article{orbansiquiera:2020,
   doi = {10.1007/s10589-020-00201-2},
@@ -111,7 +122,8 @@ journal={Technical report 767}
 
 @book{dennisschnabel:1996,
   author    = {J.E. {Dennis Jr} and R.B. Schnabel},
-  publisher = {SIAM, Philadelphia, PA, USA},
+  publisher = {SIAM},
+  address   = {Philadelphia, PA, USA},
   title     = {Numerical Methods for Unconstrained optimization and Nonlinear Equations},
   year      = {1996},
   doi       = {10.1137/1.9781611971200},
@@ -129,9 +141,10 @@ journal={Technical report 767}
 @book{hockschittkowski:1980,
   author    = {W. Hock and K. Schittkowski},
   edition   = {Second edition},
-  publisher = {Springer, Berlin, Heidelberg, Germany},
+  publisher = {Springer Berlin},
   series    = {Lecture Notes in Economics and Mathematical Systems},
   volume    = {187},
+  address   = {Heidelberg, Germany},
   title     = {Test Examples for Nonlinear Programming Codes},
   year      = {1980},
   doi       = {10.1007/978-3-642-48320-2}
@@ -140,7 +153,8 @@ journal={Technical report 767}
 @book{nocedalwright:2006,
   author    = {J. Nocedal and S.J. Wright},
   edition   = {Second edition},
-  publisher = {Springer, New York, NY, USA},
+  publisher = {Springer},
+  address   = {New York, NY, USA},
   series    = {Springer series in Operation Reasearch and Financial Engineering},
   title     = {Numerical Optimization},
   year      = {2006},
