### 2024-04-23 UncertainLab/Enlsip.jl#2 @odunbar
OK great! I like the table, thank you for the response and I'll take your word for it on the software paper typo's
