### 2024-04-24 UncertainLab/Enlsip.jl#3 @tmigot
Hi @pierre-borie ! I have more comments regarding the paper:

- [x] Typos:
- least squares OR least-squares
- missing "problem" in first sentence (and elsewhere)
- "This type of problems" plural/singular
In general, please, revise carefully the English.
- [x] Mention somewhere that the aim is to find a local minimum of the problem. I noticed it is well done in the documentation of the package.
- [x] Is Figure 1 a Dolan&Moré performance profile? Why the percentage is 100% with 26 problems failed?
- [x] The package JSOSolvers.jl has least squares variants for the solvers TRUNK (unconstrained) and TRON (bound-constrained).
- [x] I would use citation for the different packages when available instead of just naming them. You can usually find citation file or use Github "Cite this repository" on packages' Github page.
