### 2024-04-24 UncertainLab/Enlsip.jl#3 @pierre-borie
Hi @tmigot , thanks for your comments! Just to make sure I get these right:

- I deliberatly distinguished "least squares" from "least-squares". I use the latter as an adjective, kind of, like in "least-squares problem" or "least-squares structure" but I'll uniformize it to avoid any ambiguities;
- Figure 1 is indeed a "home made" Dolan-Moré performance profile; the percentage was still 100% because both versions failed on the same problems and we then excluded them from the test set, which is not very clear in the paper I agree. I'll replace the performance profile by a graph obtained with the BenchmarkProfiles.jl package and will correct the percentage issue at the same time.
