+++ paper/paper.bib
@@ -112,13 +112,33 @@ author={A. Griewank},
 year={2003},
 pages={321–398}}
 
+@article{linmore:1999a,
+	author = {C.-J. Lin and J.J Mor\'e},
+	title = {Newton's Method for Large Bound-Constrained Optimization Problems},
+	journal = {SIAM Journal on Optimization},
+	volume = {9},
+	number = {4},
+	pages = {1100-1127},
+	year = {1999},
+	doi = {10.1137/S1052623498345075}}
+	
+
 @article{lucksanvlcek:1999,
 author={L. Luk{\v s}an and J. Vl{\v c}ek},
 title={Sparse and Partially Separable Test Problems for Unconstrained and Equality Constrained Optimization},
 url={http://hdl.handle.net/11104/0123965},
 year={1999},
-journal={Technical report 767}
-}
+journal={Technical report 767}}
+
+@article{wachterbiegler:2006,
+	author = {A. W\"achter and L.T. Biegler},
+	title = {On the implementation of an interior-point filter line-search algorithm for large-scale nonlinear programming},
+	journal = {Mathematical Programming, Series A},
+	volume = {106},
+	pages = {25-57},
+	year = {2006},
+	doi = {10.1007/s10107-004-0559-y}}
+
 
 @book{dennisschnabel:1996,
   author    = {J.E. {Dennis Jr} and R.B. Schnabel},
@@ -147,8 +167,7 @@ journal={Technical report 767}
   address   = {Heidelberg, Germany},
   title     = {Test Examples for Nonlinear Programming Codes},
   year      = {1980},
-  doi       = {10.1007/978-3-642-48320-2}
-}
+  doi       = {10.1007/978-3-642-48320-2}}
 
 @book{nocedalwright:2006,
   author    = {J. Nocedal and S.J. Wright},
@@ -158,5 +177,13 @@ journal={Technical report 767}
   series    = {Springer series in Operation Reasearch and Financial Engineering},
   title     = {Numerical Optimization},
   year      = {2006},
-  doi       = {10.1007/978-0-387-40065-5}
-}
+  doi       = {10.1007/978-0-387-40065-5}}
+
+@software{jsosolvers:2023,
+author = {T. Migot and D. Orban and A.S. Siqueira and {contributors}},
+license = {MPL-2.0},
+month = jul,
+title = {{JSOSolvers.jl: JuliaSmoothOptimizers optimization solvers}},
+url = {https://github.com/JuliaSmoothOptimizers/JSOSolvers.jl},
+version = {0.11.0},
+year = {2023}}
