### 2024-04-30 UncertainLab/Enlsip.jl#5 @tmigot
Comments on README
Just a couple of comments on the README:
- [x]  In the intro, you mentioned that this package is the "Julia version of an optimization library originally written in Fortran77". Is this Fortran library available online? If not I would remove the statement as it is probably not necessary for potential users. Otherwise, could you add a link? What do you think?
- [x] It would be good to add a complete citation or a link to the original paper from Per Lindstrom and Per Ake Wedin.
- [x] norm in after the problem statement is not the same as in the problem. (I noticed it is ok in the documentation so this is probably a markdown issue)
- [x] "First, you need to define to create a model of your problem with the CnlsModel structure." -> choose "to define" or "to create"
- [x] The mathematical formulation does not include bounds constraints while you CnlsModel include those. It could be good to add them in the mathematical formulation and just precise that those are treated as classical inequalities.

https://github.com/openjournals/joss-reviews/issues/6226
