+++ paper/paper.bib
@@ -20,7 +20,7 @@
 
 @article{lindstromwedin1988,
   author   = {P. Lindstr{\"o}m and P.{\AA}. Wedin},
-  journal  = {Technical Report S-901 87 from the Institute of Information processing, University of Ume{\aa}, Sweden},
+  journal  = {Technical Report S-901 87, Institute of Information processing, University of Ume{\aa}, Sweden},
   keywords = {Nonlinear least squares, constraints, Gauss-Newton, quadratic merit function, subspace minimization},
   title    = {{Gauss-Newton} based algorithms for constrained nonlinear least squares problems},
   year     = {1988}
