### 2024-04-30 UncertainLab/Enlsip.jl#3 @pierre-borie
@tmigot - I submitted a new version of the paper that will, I hope, address your latest comments!
