### 2024-04-30 #6226 @pierre-borie
@jbytecode I pushed another revised version of the paper, correcting (hopefully) the lastest issues.

@odunbar Thanks for supporting the submission! The new version does not differ much from the previous one
 - The comparison graph have been replaced with a performance profile obtained with a dedicated package
 - Some sentences have been reformulated and citations of other least-squares packages added
