### 2024-05-01 UncertainLab/Enlsip.jl#6 @tmigot
Reproducible example with Ipopt in tutorial
In the tutorial, there is a table with numerical resutls from Ipopt.
This comparison is difficult to assess because there is no reproducible code for this example.

I would suggest to either add the code or simply remove the table. It is not clear from this table how comparable the two solves are, for instance, do they have the same stopping criteria?

https://github.com/openjournals/joss-reviews/issues/6226
