### 2024-05-01 UncertainLab/Enlsip.jl#8 @tmigot
Add tolerance as input argument
I imagine a potential user would most likely be interested in setting the stopping tolerance by himself.
Besides, I don't think the stopping criterion is explicited somewhere in the documentation

https://github.com/openjournals/joss-reviews/issues/6226
