### 2024-05-01 1b3e276
Fix some typos in tutorial
