### 2024-05-14 848247a
added stopping criteria to documentation
