+++ docs/src/method.md
@@ -1,0 +1,114 @@
+# [Method](@id Method)
+
+In this section, a brief description of the method implemented in `Enlsip.jl` is given.
+
+We recall that the problem to solve is of the form:
+
+```math
+\begin{aligned}
+\min_{x \in \mathbb{R}^n} \quad &  \dfrac{1}{2} \|r(x)\|^2 \\
+\text{s.t.} \quad & c_i(x) = 0, \quad i \in \mathcal{E} \\
+& c_i(x) \geq 0, \quad i \in \mathcal{I}, \\
+\end{aligned}
+```
+
+with:
+
+* the residuals $r_i:\mathbb{R}^n\rightarrow\mathbb{R}$ and the constraints $c_i:\mathbb{R}^n\rightarrow\mathbb{R}$ assumed to be $\mathcal{C}^2$ functions;
+* norm $\|\cdot\|$ denotes the Euclidean norm.
+
+We introduce the Lagrangian associated to the problem:
+
+```math
+L(x,\lambda) = \dfrac{1}{2} \|r(x)\|^2 - \sum_{i \in \mathcal{E} \cup \mathcal{I}} \lambda_i c_i(x),
+```
+
+where $\lambda = \left(\lambda_i\right)_{i\in \mathcal{E} \cup \mathcal{I}}$ denotes the vector of Lagrange multipliers.
+
+## Conduct of the algorithm
+
+Starting from a point $x_0$, the algorithm builds a sequence $(x_k)_k$ converging to a first order critical point of the problem.
+
+At a given iteration $k$, a search direction $p_k\in\mathbb{R}^n$ and a steplength $\alpha_k\in[0,1]$ are computed such that the next step is $x_{k+1}:=x_k+\alpha_kp_k$.
+
+An estimate $\lambda_{k}\in\mathbb{R}^{|\mathcal{E}| + |\mathcal{I}|}$ of the vector of Lagrange multipliers is then computed and convergence is tested.
+
+### Search direction
+
+The search direction is the solution of a subproblem derived from an approximation of the original problem.
+
+First, the residuals are linearized in a small neighborhood of current point $x_k$:
+
+$$r(x_k+p)\approx J_kp+r_k,$$
+
+where $J_k$ denotes the Jacobian matrix of the residuals function evaluated at $x_k$ and $r_k:=r(x_k)$ . The resulting objective function of the subproblem corresponds to the Gauss-Newton approximation.
+
+The constraints of the subproblem are formed after a subset of the constraints of the original problem. It contains all the equality constraints but only the inequality constraints that are believed to be satisfied with equality at the solution. This subset, often denoted as the working set in the literature, is updated at every iteration and can be seen as a guess of the optimal active set. For more details on the active set strategy implemented in `Enlsip.jl`, see chapters 5 and 6 of *Practical Optimization*[^GMW82].
+
+Noting $\hat{c}_k$ the vector of constraints in the working set evaluated at $x_k$ and $\hat{A}_k$ its Jacobian matrix, the subproblem is then given by:
+
+```math
+\begin{aligned}
+\min_{p \in \mathbb{R}^n} \quad &  \dfrac{1}{2} \|J_kp+r_k\|^2 \\
+\text{s.t.} \quad & \hat{A}_kp + \hat{c}_k= 0.
+\end{aligned}
+```
+
+The subproblem is then solved by a null-space type method.
+
+### Steplength
+
+The steplength aims to maintain feasibility  of all of the constraints and to reduce the value of the objective function. In `Enlsip.jl`, this process involves an $\ell_2$-type merit function:
[…]
