### 2024-05-14 a811c09
renaming objective_value method + add documentation about tolerances as input arguments + model chained rosenbrock with ipopt
