### 2024-05-14 UncertainLab/Enlsip.jl#8 @pierre-borie
Indeed, a description of the stopping criteria was clearly missing. I added one to the documentation. Tolerances can be passed as keyword arguments of the main solving function.

#10
