### 2024-05-14 UncertainLab/Enlsip.jl#6 @pierre-borie
I think it's relevant; the idea is just to show what can be expected in terms of computation times. I added the modeling code of the problem with JuMP (and its solving with ipopt). I hope it will clarify the comparison.

#10
