### 2024-05-14 #6226 @pierre-borie
@jbytecode - I created a pull request covering the last issues about the package.
