### 2024-05-14 0610098
modified Project.toml for v0.9.6 release
