+++ paper/paper.md
@@ -76,7 +76,7 @@ The performance of the two versions have been compared on problems derived from 
 
 Due to the differences in how computations are organized between Fortran77 and Julia, especially in linear algebra, minor numerical discrepancies were expected to emerge and accumulate, leading to slightly different outcomes or total number of iterations performed.
 
-Out of 90 instances, 26 either failed to converge or stopped because of a numerical error occuring during the execution of both ``ENLSIP`` versions. These results were due to inconsistencies in the problem formulation of a given family of instances and thus, they do not reflect the robustness of the method implemented in ``ENLSIP``. For the 64 remaining instances, both versions reached similar solutions that met Hydro-Québec specifications. Although specific values cannot be disclosed in this paper, the performance profile [@dolanmore:2002] in Figure \ref{fig:perf_profile} suggests that our Julia version has similar, if not better, performance than the Fortran77 version. However, these results must be weighted by the number of iterations performed by the two algorithms. Indeed, our Julia version often required less iterations. We argue that the better efficiency of the linear algebra routines from OpenBLAS as compared to those from MINPACK may contribute to shorter computation times.
+Out of 90 instances, 26 either failed to converge or stopped because of a numerical error occurring during the execution of both ``ENLSIP`` versions. These results were due to inconsistencies in the problem formulation of a given family of instances and thus, they do not reflect the robustness of the method implemented in ``ENLSIP``. For the 64 remaining instances, both versions reached similar solutions that met Hydro-Québec specifications. Although specific values cannot be disclosed in this paper, the performance profile [@dolanmore:2002] in Figure \ref{fig:perf_profile} suggests that our Julia version has similar, if not better, performance than the Fortran77 version. However, these results must be weighted by the number of iterations performed by the two algorithms. Indeed, our Julia version often required less iterations. We argue that the better efficiency of the linear algebra routines from OpenBLAS as compared to those from MINPACK may contribute to shorter computation times.
 
 ![Computation time performance profile of ENLSIP-Fortran77 (dashed curve) and Enlsip.jl (continuous curve) on 90 instances of the CM1 model calibration.\label{fig:perf_profile}](figures/performance_profile_cm1.png)
 
