+++ v0.9.6/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"method/#Method","page":"Method","title":"Method","text":"","category":"section"},{"location":"method/","page":"Method","title":"Method","text":"In this section, a brief description of the method implemented in Enlsip.jl is given.","category":"page"},{"location":"method/","page":"Method","title":"Method","text":"We recall that the problem to solve is of the form:","category":"page"},{"location":"method/","page":"Method","title":"Method","text":"beginaligned\nmin_x in mathbbR^n quad   dfrac12 r(x)^2 \ntextst quad  c_i(x) = 0 quad i in mathcalE \n c_i(x) geq 0 quad i in mathcalI \nendaligned","category":"page"},{"location":"method/","page":"Method","title":"Method","text":"with:","category":"page"},{"location":"method/","page":"Method","title":"Method","text":"the residuals r_imathbbR^nrightarrowmathbbR and the constraints c_imathbbR^nrightarrowmathbbR assumed to be mathcalC^2 functions;\nnorm cdot denotes the Euclidean norm.","category":"page"},{"location":"method/","page":"Method","title":"Method","text":"We introduce the Lagrangian associated to the problem:","category":"page"},{"location":"method/","page":"Method","title":"Method","text":"L(xlambda) = dfrac12 r(x)^2 - sum_i in mathcalE cup mathcalI lambda_i c_i(x)","category":"page"},{"location":"method/","page":"Method","title":"Method","text":"where lambda = left(lambda_iright)_iin mathcalE cup mathcalI denotes the vector of Lagrange multipliers.","category":"page"},{"location":"method/#Conduct-of-the-algorithm","page":"Method","title":"Conduct of the algorithm","text":"","category":"section"},{"location":"method/","page":"Method","title":"Method","text":"Starting from a point x_0, the algorithm builds a sequence (x_k)_k converging to a first order critical point of the problem.","category":"page"},{"location":"method/","page":"Method","title":"Method","text":"At a given iteration k, a search direction p_kinmathbbR^n and a steplength alpha_kin01 are computed such that the next step is x_k+1=x_k+alpha_kp_k.","category":"page"},{"location":"method/","page":"Method","title":"Method","text":"An estimate lambda_kinmathbbR^mathcalE + mathcalI of the vector of Lagrange multipliers is then computed and convergence is tested.","category":"page"},{"location":"method/#Search-direction","page":"Method","title":"Search direction","text":"","category":"section"},{"location":"method/","page":"Method","title":"Method","text":"The search direction is the solution of a subproblem derived from an approximation of the original problem.","category":"page"},{"location":"method/","page":"Method","title":"Method","text":"First, the residuals are linearized in a small neighborhood of current point x_k:","category":"page"},{"location":"method/","page":"Method","title":"Method","text":"r(x_k+p)approx J_kp+r_k","category":"page"},{"location":"method/","page":"Method","title":"Method","text":"where J_k denotes the Jacobian matrix of the residuals function evaluated at x_k and r_k=r(x_k) . The resulting objective function of the subproblem corresponds to the Gauss-Newton approximation.","category":"page"},{"location":"method/","page":"Method","title":"Method","text":"The constraints of the subproblem are formed after a subset of the constraints of the original problem. It contains all the equality constraints but only the inequality constraints that are believed to be satisfied with equality at the solution. This subset, often denoted as the working set in the literature, is updated at every iteration and can be seen as a guess of the optimal active set. For more details on the active set strategy implemented in Enlsip.jl, see chapters 5 and 6 of Practical Optimization[GMW82].","category":"page"},{"location":"m
[…]
