### 2024-05-14 #6226 @tmigot
Hi @jbytecode

With the last PR now merged, @pierre-borie answered all my comments and I completed the check-list.
I am happy to recommend the publication of Enlsip in JOSS.

The nonlinear least squares problem is both important and very frequent in practice, so I think this algorithm will be an interesting addition to the Julia background. My only comment would be that modeling a problem for the solver is not an easy task. I would recommend relying on NLPModels.jl to interact with the model as it would open the door to several modeling and benchmarking tools, but that's a biased opinion and independent of this submission process.

Congrats @pierre-borie for the nice work!
