### 2024-05-15 #6226 @tmigot
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1109/PES.2006.1709029 is OK
> - 10.1007/BF02591997 is OK
> - 10.1137/141000671 is OK
> - 10.1137/15M1020575 is OK
> - 10.5281/zenodo.2558627 is OK
> - 10.1145/355958.355965 is OK
> - 10.1007/s101070100263 is OK
> - 10.1007/s10589-020-00201-2 is OK
> - 10.1017/S0962492902000132 is OK
> - 10.1137/S1052623498345075 is OK
> - 10.1007/s10107-004-0559-y is OK
> - 10.1137/1.9781611971200 is OK
> - 10.1007/978-0-387-40065-5 is OK
> - 10.1007/978-3-642-48320-2 is OK
> - 10.1007/978-0-387-40065-5 is OK
>
> MISSING DOIs
>
> - No DOI given, and none found for title: Gauss-Newton based algorithms for constrained nonl...
> - No DOI given, and none found for title: Sparse and Partially Separable Test Problems for U...
> - No DOI given, and none found for title: JSOSolvers.jl: JuliaSmoothOptimizers optimization ...
>
> INVALID DOIs
>
> - None
> ```

DOI for JSOSolvers: https://github.com/JuliaSmoothOptimizers/JSOSolvers.jl/blob/main/CITATION.cff
