+++ paper/paper.bib
@@ -122,13 +122,12 @@ pages={321–398}}
 	year = {1999},
 	doi = {10.1137/S1052623498345075}}
 	
-
 @article{lucksanvlcek:1999,
 author={L. Luk{\v s}an and J. Vl{\v c}ek},
 title={Sparse and Partially Separable Test Problems for Unconstrained and Equality Constrained Optimization},
 url={http://hdl.handle.net/11104/0123965},
 year={1999},
-journal={Technical report 767}}
+journal={Technical report 767, Institute of Computer Science, Academy of Sciences of the Czech Republic, Prague}}
 
 @article{wachterbiegler:2006,
 	author = {A. W\"achter and L.T. Biegler},
@@ -182,7 +181,7 @@ journal={Technical report 767}}
 @software{jsosolvers:2023,
 author = {T. Migot and D. Orban and A.S. Siqueira and {contributors}},
 license = {MPL-2.0},
-month = jul,
+doi = {10.5281/zenodo.3991143},
 title = {{JSOSolvers.jl: JuliaSmoothOptimizers optimization solvers}},
 url = {https://github.com/JuliaSmoothOptimizers/JSOSolvers.jl},
 version = {0.11.0},
