### 2024-05-15 #6226 @pierre-borie
@jbytecode - I added the DOI for JSOSolvers. For the other two papers, I could not find any. I also checked in papers/books citing them and they are referenced without DOI so I believe they don't have one (perhaps because they are technical reports).
