### 2024-05-15 #6226 @pierre-borie
@tmigot - Thanks for approving the corrections and for your advices! The current modeling interface isn't the best one indeed and switching NLPModels.jl is definitely something I intend to work on.
