### 2024-05-16 #6226 @jbytecode
@pierre-borie - I have just had a read of the manuscript. Thank you, good work. I've sent a pull request that includes some minor changes on the manuscript. Could you please review them? If you are okay with them, please merge. If you see a problem, please tell me.

- Some of the authors' ORCIDs are mission. Please provide ORCIDs for all of the authors.
- In Equation 2, is it better to use `argmax` rather than a single `max` just because you provide names of the decision variables under the `max` keyword?

Thank you in advance.
