### 2024-05-16 #6226 @pierre-borie
@jbytecode thank you for your comment, I am okay with your remarks and merged your pull request.

>     * In Equation 2, is it better to use `argmax` rather than a single `max` just because you provide names of the decision variables under the `max` keyword?

In this context the emphasis is on the objective function and not the decision variables per se so I think it should be a `min` rather than `argmin` but I understand your point. Would it make it less ambiguous to replace the `x_1, x_2, x_3` under `min` by simply `x`?

>     * Some of the authors' ORCIDs are missing. Please provide ORCIDs for all of the authors.

Is it mandatory for all the authors to provide an ORCID? The two authors concerned don't have one (but it should not be a problem for them to create one if necessary).

Once that's sorted out, I will take care of the release and archiving and come back to you when it's done.
