### 2024-05-16 #6226 @jbytecode
- Yes, perhaps it is better to remove $x_1$, $x_2$, and $x_3$ and go on.
- Please, provide the ORCID.

After performing these task, please create a tagged release in the repo. Using this tagged release create a Zenodo archive. Please make sure that the title, author list and ORCIDs match with those in the paper and submission.

Thank you in advance.
