### 2024-05-16 365a6ac
modified problem 65 formulation in paper.md
