### 2024-05-21 #6226 @pierre-borie
@jbytecode - Hi, I added the missing DOI to the paper, made a tagged release and an archive.

- The version tag of the JOSS release is [v0.9.7](https://github.com/UncertainLab/Enlsip.jl/releases/tag/v0.9.7)
- Archive DOI is [10.5281/zenodo.11206280](https://doi.org/10.5281/zenodo.11206280) (license in the archive is the same as the software license)
- I also checked that the title, author list and ORCIDs match with those in the paper.

Let me know if there are any issues!
