### 2024-05-22 #6226 @jbytecode
@openjournals/dev - Hi, sorry for bothering, what is this error about? I couldn't address the problem. the editorialbot compiles the pdf correctly but we get this error after acceptance. Could you please help us? Thank you!
