### 2024-05-22 #6226 @xuanxu
The errors come from the math equations code using displaymode with `\dfrac`. It looks like that function is not fully supported by our compilation tool. [This PR](https://github.com/UncertainLab/Enlsip.jl/pull/13) modifies the code for the equations, and running `recommend-accept` after is merged should not display those errors.
