### 2024-05-22 #6226 @crvernon
:wave:  @pierre-borie - this submission is looking really good!  I only have the one issue for you to see to before I accept this for publication.

- The DOI in the following reference is pointing to the incorrect location:

> Gill, P. E., Murray, W., & Wright, M. H. (1981). Practical optimization. Academic Press. https://doi.org/10.1007/978-0-387-40065-5

Instead it is pointing to the same DOI as this reference:

> Nocedal, J., & Wright, S. J. (2006). Numerical optimization (Second edition). Springer. https://doi.org/10.1007/978-0-387-40065-5

Let me know when you have fixed this and we will move forward.
