### 2024-05-22 5c597bd
Update paper.bib
Corrected Practical optimization doi
