+++ paper/paper.bib
@@ -155,7 +155,7 @@ journal={Technical report 767, Institute of Computer Science, Academy of Science
  	address   = {San Diego, CA, USA}, 
  	title     = {Practical Optimization},
  	year      = {1981},
- 	doi       = {10.1007/978-0-387-40065-5}}
+ 	doi       = {10.1137/1.9781611975604}}
 
 @book{hockschittkowski:1980,
   author    = {W. Hock and K. Schittkowski},
