### 2024-05-22 #6226 @pierre-borie
Thanks a lot @jbytecode for the editing and @tmigot @odunbar for the reviews. The package and its documentation has clearly benefited from your comments!
